Offset by 4 error!
Problem is that I defined the header's reserved and handles fields as
4 bytes (unsigned int) instead of 2 bytes each (unsigned short)... oops

+++ include/header.hxx
@@ -84,8 +84,8 @@ typedef struct {
     unsigned int    version;
     unsigned int    bytes;
     unsigned int    records;
-    unsigned int    handles;
-    unsigned int    reserved;
+    unsigned short  handles;
+    unsigned short  reserved;
     unsigned int    nDescription;
     unsigned int    offDescription;
     unsigned int    nPalEntries;
